test: re-organize tests
